+++ src/aui-io/js/aui-io-request.js
@@ -294,6 +294,9 @@ var IORequest = A.Component.create({
          * @type String
          */
         method: {
+            setter: function(val) {
+                return val.toLowerCase();
+            },
             valueFn: getDefault('method')
         },