+++ src/aui-pagination/tests/unit/js/tests.js
@@ -40,6 +40,30 @@ YUI.add('aui-pagination-tests', function(Y) {
         }
     }));
 
+    //--------------------------------------------------------------------------
+    // Test Case for Testing Pagination Links States
+    //--------------------------------------------------------------------------
+
+    suite.add(new Y.Test.Case({
+
+        name: 'Pagination link state tests',
+
+        // Tests: AUI-1274
+        'Assert that Pagination State and Pagination UI are in Sync': function() {
+            var paginationItems = Y.one('#pagination .pagination-content').all('li'),
+                state = {
+                    page: 5
+                },
+                pageIndex = (state.page - 1);
+
+            pagination.setState(state);
+
+            paginationItems.each(function(item, index) {
+                Y.Test.Assert.isTrue((index === pageIndex) === item.hasClass('active'));
+            });
+        }
+    }));
+
     Y.Test.Runner.add(suite);
 
 }, '', {