+++ src/aui-video/js/aui-video.js
@@ -108,6 +108,18 @@ var Video = A.Component.create({
         },
 
         /**
+         * The required Flash version for the swf player
+         *
+         * @attribute flashPlayerVersion
+         * @default '9,0,0,0'
+         * @type String
+         */
+        flashPlayerVersion: {
+            validator: Lang.isString,
+            value: '9,0,0,0'
+        },
+
+        /**
          * Variables used by Flash player.
          *
          * @attribute flashVars
@@ -277,7 +289,9 @@ var Video = A.Component.create({
                 var tplObj = '<object id="' + instance._swfId + '" ';
 
                 if (UA.ie) {
-                    tplObj += 'classid="clsid:d27cdb6e-ae6d-11cf-96b8-444553540000" ';
+                    tplObj += 'classid="clsid:d27cdb6e-ae6d-11cf-96b8-444553540000" ' +
+                        'codebase="http://download.macromedia.com/pub/shockwave/cabs/flash/swflash.cab#version=' +
+                        instance.get('flashPlayerVersion') + '" ';
                 }
                 else {
                     tplObj += 'type="application/x-shockwave-flash" data="' + swfUrl + '" ';