+++ src/aui-io/HISTORY.md
@@ -9,3 +9,4 @@ No registries yet.
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)
 
 * [AUI-1163](https://issues.liferay.com/browse/AUI-1163) Remove unnecessary constants
+* [AUI-843](https://issues.liferay.com/browse/AUI-843) IO-Request module should accept either GET or get